modules/kafka-location-consumer/requirements.txt
marshmallow==3.26.*
kafka-python==2.1.*
grpcio-tools==1.71.*